refactor: 강사님 코드리뷰 3차 수정(타이머 시간 변경 로직 utils로 분리)
